requirements.txt
Brotli==1.0.9
certifi==2023.5.7
ffmpeg-python==0.2.0
future==0.18.3
mutagen==1.46.0
pycryptodomex==3.18.0
websockets==11.0.3
yt-dlp==2023.6.22
